refactor: 책임 연쇄 패턴을 적용한 로직으로 변경
