--- src/main/java/nextstep/subway/path/domain/discount/DiscountPolicy.java
@@ -1,5 +1,0 @@
-package nextstep.subway.path.domain.discount;
-
-public interface DiscountPolicy {
-    int discount(int totalFare);
-}
