--- src/main/java/nextstep/subway/path/domain/discount/age/AgeDiscountPolicy.java
@@ -1,7 +1,0 @@
-package nextstep.subway.path.domain.discount.age;
-
-import nextstep.subway.path.domain.discount.DiscountPolicy;
-
-public abstract class AgeDiscountPolicy implements DiscountPolicy {
-    protected static final int DEDUCTION_AMOUNT = 350;
-}
