--- src/main/java/nextstep/subway/path/domain/fare/distance/DistanceFarePolicy.java
@@ -1,7 +1,0 @@
-package nextstep.subway.path.domain.fare.distance;
-
-public interface DistanceFarePolicy {
-    boolean isSatisfiedBy(int totalDistance);
-
-    int calculateFare(int totalDistance);
-}
