--- src/main/java/nextstep/subway/path/domain/fare/distance/MiddleDistanceFarePolicy.java
@@ -1,18 +1,0 @@
-package nextstep.subway.path.domain.fare.distance;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class MiddleDistanceFarePolicy extends DefaultDistanceFarePolicy {
-    @Override
-    public boolean isSatisfiedBy(int totalDistance) {
-        return SHORT_DISTANCE_LIMIT < totalDistance
-                && totalDistance <= MIDDLE_DISTANCE_LIMIT;
-    }
-
-    @Override
-    public int calculateFare(int totalDistance) {
-        int lastDistance = totalDistance - SHORT_DISTANCE_LIMIT;
-        return BASIC_FEE + calculateOverFare(lastDistance, MIDDLE_DISTANCE_UNIT);
-    }
-}
