--- src/main/java/nextstep/subway/path/domain/fare/line/LineFarePolicy.java
@@ -1,7 +1,0 @@
-package nextstep.subway.path.domain.fare.line;
-
-import nextstep.subway.line.domain.Lines;
-
-public interface LineFarePolicy {
-    int calculateAdditionalFare(Lines lines);
-}
