+++ src/main/java/nextstep/subway/path/domain/policy/discount/DefaultAgeDiscountPolicy.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain.discount.age;
+package nextstep.subway.path.domain.policy.discount;
 
 public class DefaultAgeDiscountPolicy extends AgeDiscountPolicy {
     @Override
