+++ src/main/java/nextstep/subway/path/domain/policy/discount/DefaultDiscountPolicy.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.domain.discount;
+package nextstep.subway.path.domain.policy.discount;
 
 public class DefaultDiscountPolicy implements DiscountPolicy {
     @Override
