+++ src/main/java/nextstep/subway/path/domain/policy/discount/DiscountPolicy.java
@@ -1,4 +1,5 @@
 package nextstep.subway.path.domain.policy.discount;
 
 public interface DiscountPolicy {
+    int discount(int totalFare);
 }
